rename, update tests, fix types
