+++ src/wallet/components/WalletIsland.tsx
